+++ resources/js/views/admin/nivel/index.vue
@@ -28,10 +28,9 @@
                                 </svg>
                             </router-link>
                             <!--Botón para eliminar nivel-->
-                            <Toast />
+                        
 
-                            <ConfirmPopup> </ConfirmPopup>
-                            <Button @click="confirm1($event,slotProps.data.id, slotProps.index)"  class="btn btn-primary">
+                            <Button  class="btn btn-primary" @click="confirm1($event,slotProps.data.id, slotProps.index)">
                                 <svg xmlns="http://www.w3.org/2000/svg" height="1em" viewBox="0 0 448 512">
                                     <path fill="#ffffff"
                                         d="M135.2 17.7C140.6 6.8 151.7 0 163.8 0H284.2c12.1 0 23.2 6.8 28.6 17.7L320 32h96c17.7 0 32 14.3 32 32s-14.3 32-32 32H32C14.3 96 0 81.7 0 64S14.3 32 32 32h96l7.2-14.3zM32 128H416V448c0 35.3-28.7 64-64 64H96c-35.3 0-64-28.7-64-64V128zm96 64c-8.8 0-16 7.2-16 16V432c0 8.8 7.2 16 16 16s16-7.2 16-16V208c0-8.8-7.2-16-16-16zm96 0c-8.8 0-16 7.2-16 16V432c0 8.8 7.2 16 16 16s16-7.2 16-16V208c0-8.8-7.2-16-16-16zm96 0c-8.8 0-16 7.2-16 16V432c0 8.8 7.2 16 16 16s16-7.2 16-16V208c0-8.8-7.2-16-16-16z" />
@@ -50,15 +49,9 @@
 
 <script setup>
   
-  import {ref, onMounted} from "vue"
-
+  import {ref, onMounted,inject} from "vue"
+  const swal = inject('$swal')
   const niveles=ref()
-
-  import { useConfirm } from "primevue/useconfirm";
-  import { useToast } from "primevue/usetoast";
-
-  const confirm = useConfirm();
-  const toast = useToast();
 
   onMounted(()=>{
 
@@ -85,28 +78,35 @@
             });
     };
 
-    //Función para popup
+    //Función para mensaje confirmación
     const confirm1 = (event,id,index) => {
         console.log(event);
+        swal({
+            title: 'Estás seguro??',
+            text: 'No podrás revertir esta acción!',
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonText: 'Si',
+            confirmButtonColor: '#ef4444',
+            timer: 20000,
+            timerProgressBar: true,
+            reverseButtons: true
+        })
+            .then(result => {
+                if (result.isConfirmed) {
+                    swal({
+                                icon: 'success',
+                                title: 'Dato eliminado'
+                            })
+                    deleteNivel(id,index)
+                }else{
+                    swal({
+                                icon: 'error',
+                                title: 'Error al intentar eliminar el dato'
+                            })
+                }
+            })
 
-        confirm.require({
-            target: event.currentTarget,
-            message: 'Estas seguro que deseas eliminar este dato?',
-            icon: 'pi pi-exclamation-triangle',
-            rejectClass: 'p-button-secondary p-button-outlined p-button-sm',
-            acceptClass: 'p-button-sm',
-            rejectLabel: 'No',
-            acceptLabel: 'Si',
-            accept: () => {
-                toast.add({ severity: 'info', summary: 'Confirmado', detail: 'Dato eliminado', life: 3000 });
-                deleteNivel(id,index)
-            },
-            reject: () => {
-
-            
-                toast.add({ severity: 'error', summary: 'Cancelado', detail: 'Cambios no guardados', life: 3000 });
-            }
-        });
     };
   
 </script>
